Cleaned up buildRecentModelRevisions file
